refactor: LottoGenerator역할 Lotto에 위임
