+++ src/test/java/lotto/domain/LottosTest.java
@@ -15,16 +15,16 @@ public class LottosTest {
     @BeforeEach
     void setUp() {
         lottoGeneratorService = new LottoGeneratorService();
-        wonLotto = new WonLotto(lottoGeneratorService.lottoStringParser("1,2,3,4,5,6"), new LottoNumber(7));
+        wonLotto = new WonLotto(new Lotto("1,2,3,4,5,6"), new LottoNumber(7));
     }
 
     @Test
     void testLottoPrize() {
         Lottos lottos = new Lottos();
-        lottos.add(lottoGeneratorService.lottoStringParser("1,2,3,4,5,6"));
-        lottos.add(lottoGeneratorService.lottoStringParser("1,2,3,4,5,7"));
-        lottos.add(lottoGeneratorService.lottoStringParser("1,2,3,4,5,10"));
-        lottos.add(lottoGeneratorService.lottoStringParser("7,8,9,10,11,12"));
+        lottos.add(new Lotto("1,2,3,4,5,6"));
+        lottos.add(new Lotto("1,2,3,4,5,7"));
+        lottos.add(new Lotto("1,2,3,4,5,10"));
+        lottos.add(new Lotto("7,8,9,10,11,12"));
         assertThat(Arrays.asList(LottoRank.FIRST, LottoRank.SECOND, LottoRank.THIRD, LottoRank.NONE)).isEqualTo(lottos.lottosResult(wonLotto));
     }
